Move MetadataStorage to it's own header.
